@@ -18,14 +18,14 @@
 	<packaging>pom</packaging>
 
 	<properties>
-		<java.version>8</java.version>
+		<java.version>11</java.version>
 	</properties>
 
 
 
 	<modules>
 		<module>java-basic</module>
-		<module>frontend</module>
+<!--		<module>frontend</module>-->
 		<module>backend</module>
 		<module>java-service</module>
 	</modules>
